+++ src/utils-win.h
@@ -27,13 +27,12 @@
  * SUCH DAMAGE.
  */
 
-
 #ifndef _RIFIUTI_UTILS_WIN_H
 #define _RIFIUTI_UTILS_WIN_H
 
 void       gui_message              (const char          *message         );
-
 char *     get_win32_locale         (void);
+_Bool      can_list_win32_folder    (const char          *dir             );
 
 #endif